+++ src/bittytax/conv/mergers/__init__.py
@@ -1,1 +1,1 @@
-from . import bscscan, etherscan, ftmscan, hecoinfo, polygonscan, snowtrace
+from . import bscscan, coincorner, etherscan, ftmscan, hecoinfo, polygonscan, snowtrace