+++ src/bittytax/conv/parsers/ftmscan.py
@@ -97,10 +97,10 @@ ftm_txns = DataParser(
     ParserType.EXPLORER,
     "FTMScan (FTM Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        "DateTime (UTC)",
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "From",
         "To",
         "ContractAddress",
@@ -122,10 +122,10 @@ DataParser(
     ParserType.EXPLORER,
     "FTMScan (FTM Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        "DateTime (UTC)",
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "From",
         "To",
         "ContractAddress",
@@ -148,10 +148,10 @@ ftm_int = DataParser(
     ParserType.EXPLORER,
     "FTMScan (FTM Internal Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        "DateTime (UTC)",
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "ParentTxFrom",
         "ParentTxTo",
         "ParentTxFTM_Value",
@@ -174,10 +174,10 @@ DataParser(
     ParserType.EXPLORER,
     "FTMScan (FTM Internal Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        "DateTime (UTC)",
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "ParentTxFrom",
         "ParentTxTo",
         "ParentTxFTM_Value",