+++ src/bittytax/conv/parsers/qtwallet.py
@@ -4,7 +4,7 @@
 import re
 import sys
 from decimal import Decimal
-from typing import TYPE_CHECKING, Tuple
+from typing import TYPE_CHECKING, List, Tuple
 
 from colorama import Fore
 from typing_extensions import Unpack
@@ -12,8 +12,8 @@ from typing_extensions import Unpack
 from ...bt_types import TrType
 from ...config import config
 from ...constants import WARNING
-from ..dataparser import DataParser, ParserArgs, ParserType
-from ..exceptions import UnexpectedTypeError, UnknownCryptoassetError
+from ..dataparser import ConsolidateType, DataParser, ParserArgs, ParserType
+from ..exceptions import DataRowError, UnexpectedTypeError, UnknownCryptoassetError
 from ..out_record import TransactionOutRecord
 
 if TYPE_CHECKING:
@@ -22,29 +22,87 @@ if TYPE_CHECKING:
 WALLET = "Qt Wallet"
 
 
-def parse_qt_wallet(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[ParserArgs]) -> None:
+def parse_qt_wallet(
+    data_rows: List["DataRow"], parser: DataParser, **kwargs: Unpack[ParserArgs]
+) -> None:
+
+    if parser.args:
+        amount_hdr = parser.args[0].group(1)
+    else:
+        amount_hdr = "Amount"
+
+    if parser.args and parser.args[0].group(2):
+        # Symbol in header
+        symbol = parser.args[0].group(2)
+    elif kwargs["cryptoasset"]:
+        # Symbol from argument
+        symbol = kwargs["cryptoasset"]
+    else:
+        symbol = ""
+
+    unconfirmed_include = bool(kwargs["unconfirmed"])
+
+    for data_row in data_rows:
+        if config.debug:
+            if parser.in_header_row_num is None:
+                raise RuntimeError("Missing in_header_row_num")
+
+            sys.stderr.write(
+                f"{Fore.YELLOW}conv: "
+                f" row[{parser.in_header_row_num + data_row.line_num}] {data_row}\n"
+            )
+
+        if data_row.parsed:
+            continue
+
+        try:
+            symbol, unconfirmed_include = _parse_qt_wallet_row(
+                data_row, parser, amount_hdr, symbol, unconfirmed_include, **kwargs
+            )
+        except DataRowError as e:
+            data_row.failure = e
+        except (ValueError, ArithmeticError) as e:
+            if config.debug:
+                raise
+
+            data_row.failure = e
+
+
+def _parse_qt_wallet_row(
+    data_row: "DataRow",
+    parser: DataParser,
+    amount_hdr: str,
+    symbol: str,
+    unconfirmed_include: bool,
+    **kwargs: Unpack[ParserArgs],
+) -> Tuple[str, bool]:
     row_dict = data_row.row_dict
     data_row.timestamp = DataParser.parse_timestamp(row_dict["Date"], tz=config.local_timezone)
 
-    amount, symbol = _get_amount(data_row.row[5])
+    amount, amount_symbol = _get_amount(row_dict[amount_hdr])
 
-    if parser.args and parser.args[0].group(1):
-        symbol = parser.args[0].group(1)
+    if amount_symbol:
+        # Amount has symbol
+        symbol = amount_symbol
     elif not symbol:
-        if kwargs["cryptoasset"]:
-            symbol = kwargs["cryptoasset"]
-        else:
+        sys.stderr.write(f"{WARNING} Cryptoasset cannot be identified\n")
+        sys.stderr.write(f"{Fore.RESET}Enter symbol: ")
+        symbol = input()
+        if not symbol:
             raise UnknownCryptoassetError(kwargs["filename"], kwargs.get("worksheet", ""))
 
-    if row_dict["Confirmed"] == "false" and not kwargs["unconfirmed"]:
+    if row_dict["Confirmed"] == "false" and not unconfirmed_include:
         if parser.in_header_row_num is None:
             raise RuntimeError("Missing in_header_row_num")
 
         sys.stderr.write(
             f"{Fore.YELLOW}row[{parser.in_header_row_num + data_row.line_num}] {data_row}\n"
-            f"{WARNING} Skipping unconfirmed transaction, use the [-uc] option to include it\n"
+            f"{WARNING} This transaction is unconfirmed\n"
         )
-        return
+        sys.stderr.write(f"{Fore.RESET}Include unconfirmed transactions? [y/N] ")
+        unconfirmed_include = bool(input().lower() == "y")
+        if not unconfirmed_include:
+            return symbol, unconfirmed_include
 
     if row_dict["Type"] == "Received with":
         data_row.t_record = TransactionOutRecord(
@@ -52,7 +110,7 @@ def parse_qt_wallet(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[Pa
             data_row.timestamp,
             buy_quantity=amount,
             buy_asset=symbol,
-            wallet=WALLET,
+            wallet=_get_wallet(symbol),
             note=row_dict["Label"],
         )
     elif row_dict["Type"] == "Sent to":
@@ -61,7 +119,7 @@ def parse_qt_wallet(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[Pa
             data_row.timestamp,
             sell_quantity=amount,
             sell_asset=symbol,
-            wallet=WALLET,
+            wallet=_get_wallet(symbol),
             note=row_dict["Label"],
         )
     elif row_dict["Type"] == "Mined":
@@ -70,7 +128,7 @@ def parse_qt_wallet(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[Pa
             data_row.timestamp,
             buy_quantity=amount,
             buy_asset=symbol,
-            wallet=WALLET,
+            wallet=_get_wallet(symbol),
             note=row_dict["Label"],
         )
     elif row_dict["Type"] == "Masternode Reward":
@@ -79,7 +137,7 @@ def parse_qt_wallet(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[Pa
             data_row.timestamp,
             buy_quantity=amount,
             buy_asset=symbol,
-            wallet=WALLET,
+            wallet=_get_wallet(symbol),
             note=row_dict["Label"],
         )
     elif row_dict["Type"] == "Payment to yourself":
@@ -90,7 +148,7 @@ def parse_qt_wallet(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[Pa
             sell_asset=symbol,
             fee_quantity=amount,
             fee_asset=symbol,
-            wallet=WALLET,
+            wallet=_get_wallet(symbol),
             note=row_dict["Label"],
         )
     elif row_dict["Type"] == "Name operation":
@@ -99,11 +157,13 @@ def parse_qt_wallet(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[Pa
             data_row.timestamp,
             sell_quantity=amount,
             sell_asset=symbol,
-            wallet=WALLET,
+            wallet=_get_wallet(symbol),
             note=row_dict["Label"],
         )
     else:
         raise UnexpectedTypeError(parser.in_header.index("Type"), "Type", row_dict["Type"])
+
+    return symbol, unconfirmed_include
 
 
 def _get_amount(amount: str) -> Tuple[Decimal, str]:
@@ -116,35 +176,40 @@ def _get_amount(amount: str) -> Tuple[Decimal, str]:
     return abs(Decimal(amount)), ""
 
 
+def _get_wallet(symbol: str) -> str:
+    return f"{WALLET} ({symbol})"
+
+
 def parse_vericoin_qt_wallet(
     data_row: "DataRow", parser: DataParser, **_kwargs: Unpack[ParserArgs]
 ) -> None:
     row_dict = data_row.row_dict
     data_row.timestamp = DataParser.parse_timestamp(row_dict["Date/Time"], tz=config.local_timezone)
+    symbol = "VRC"
 
     if row_dict["Type"] == "Receive":
         data_row.t_record = TransactionOutRecord(
             TrType.DEPOSIT,
             data_row.timestamp,
             buy_quantity=Decimal(row_dict["Amount"]),
-            buy_asset="VRC",
-            wallet=WALLET,
+            buy_asset=symbol,
+            wallet=_get_wallet(symbol),
         )
     elif row_dict["Type"] == "Send":
         data_row.t_record = TransactionOutRecord(
             TrType.WITHDRAWAL,
             data_row.timestamp,
             sell_quantity=abs(Decimal(row_dict["Amount"])),
-            sell_asset="VRC",
-            wallet=WALLET,
+            sell_asset=symbol,
+            wallet=_get_wallet(symbol),
         )
     elif row_dict["Type"] == "Stake":
         data_row.t_record = TransactionOutRecord(
             TrType.STAKING,
             data_row.timestamp,
             buy_quantity=Decimal(row_dict["Amount"]),
-            buy_asset="VRC",
-            wallet=WALLET,
+            buy_asset=symbol,
+            wallet=_get_wallet(symbol),
         )
     else:
         raise UnexpectedTypeError(parser.in_header.index("Type"), "Type", row_dict["Type"])
@@ -159,11 +224,12 @@ DataParser(
         "Type",
         "Label",
         "Address",
-        lambda c: re.match(r"Amount \((\w+)\)?", c),
+        lambda c: re.match(r"(Amount \((\w+)\)?)", c),
         "ID",
     ],
     worksheet_name="Qt Wallet",
-    row_handler=parse_qt_wallet,
+    all_handler=parse_qt_wallet,
+    consolidate_type=ConsolidateType.HEADER_MATCH,
 )
 
 DataParser(
@@ -171,7 +237,7 @@ DataParser(
     "Qt Wallet (i.e. Bitcoin Core, etc)",
     ["Confirmed", "Date", "Type", "Label", "Address", "Amount", "ID"],
     worksheet_name="Qt Wallet",
-    row_handler=parse_qt_wallet,
+    all_handler=parse_qt_wallet,
 )
 
 DataParser(